Refactor and split sensor remote control
