+++ custom_components/candy/const.py
@@ -8,6 +8,8 @@ DATA_KEY_COORDINATOR = "coordinator"
 CONF_INTEGRATION_TITLE = "Candy"
 CONF_KEY_USE_ENCRYPTION = "use_encryption"
 
+UNIQUE_ID_REMOTE_CONTROL = "{0}-remote_control"
+
 UNIQUE_ID_WASHING_MACHINE = "{0}-washing_machine"
 UNIQUE_ID_WASH_CYCLE_STATUS = "{0}-wash_cycle_status"
 UNIQUE_ID_WASH_REMAINING_TIME = "{0}-wash_remaining_time"
